+++ package.js
@@ -3,26 +3,23 @@
 Package.describe({
   name: "aldeed:collection2",
   summary: "Automatic validation of insert and update operations on the client and server.",
-  version: "2.5.0",
+  version: "2.8.0",
   git: "https://github.com/aldeed/meteor-collection2.git"
 });
 
 Package.onUse(function(api) {
-
-  api.use('aldeed:simple-schema@1.3.2');
-  api.imply('aldeed:simple-schema');
-
-  api.use('underscore@1.0.0');
-  api.use('check@1.0.0');
-  api.use('mongo@1.0.4');
-  api.imply('mongo');
-  api.use('minimongo@1.0.0');
-  api.use('ejson@1.0.0');
-
-  // Allow us to detect 'insecure'.
-  api.use('insecure@1.0.0', {weak: true});
+  // Automatically include all packages for now
+  api.use([
+    'aldeed:collection2-core@1.0.0',
+    'aldeed:schema-index@1.0.1',
+    'aldeed:schema-deny@1.0.1',
+  ]);
   
-  api.addFiles(['collection2.js']);
+  api.imply([
+    'aldeed:collection2-core',
+    'aldeed:schema-index',
+    'aldeed:schema-deny',
+  ]);
 });
 
 Package.onTest(function(api) {
@@ -42,6 +39,8 @@ Package.onTest(function(api) {
     'tests/collections.js',
     'tests/pubsub.js',
     'tests/security.js',
-    'tests/tests.js'
+    'tests/tests.js',
+    'tests/tests-deny.js',
+    'tests/tests-indexing.js'
   ]);
 });